refactor: rename util folder to utils (#2682)
