+++ main.ts
@@ -8,8 +8,8 @@
 
 import { ServerContext } from "$fresh/server.ts";
 import { serve } from "$std/http/server.ts";
-import { lookupSymbol } from "./util/doc_utils.ts";
-import { withLog } from "./util/ga_utils.ts";
+import { lookupSymbol } from "./utils/doc_utils.ts";
+import { withLog } from "./utils/ga_utils.ts";
 import { setup } from "$doc_components/services.ts";
 
 import twindPlugin from "$fresh/plugins/twind.ts";
