+++ utils/manual_utils.ts
