+++ utils/registry_utils.ts
