+++ utils/registry_utils_test.ts
