+++ utils/search_insights_utils.ts
